build: parametrize python & django-admin.py
This is particularly useful when cross-building Seafile in build-systems
such as OpenWrt or buildroot, where the `python` executable is not in PATH.

Manipulating the PATH variable is also doable but, it's better to provide
the full/absolute path to python, as well as to the django-admin.py script.

This change allows python & django-admin.py to be specified at a different
location, as well as maintaining their default values if unspecified.

Strictly speaking for cross-building, this change is sufficient.

Signed-off-by: Alexandru Ardelean <ardeleanalex@gmail.com>