Enable single sign out when using OpenID Connect
This allows the administrator to add the setting OAUTH_END_SESSION_URL. If they do, when users log out, Seahub will try to log them out of their SSO session globally. This is a standard feature of OpenID Connect (so strictly speaking it’s not OAuth). See the draft for session management [1] (technically it’s a draft and not yet a standard, but it is already deployed in OIDC software).

[1] https://openid.net/specs/openid-connect-session-1_0-28.html